requirements.txt
pandas==1.1.5
absl-py==0.11.0
six==1.15
numpy==1.19.2
matplotlib==3.3.3
torch==1.10.2
tensorflow-estimator==2.4.0
typing-extensions==3.7.4
protobuf==3.13.0
tensorflow-gpu==2.4.1
tensorflow-quantum==0.5.1
scikit-image==0.17.2
scikit-learn
cirq==0.11.0
sympy==1.5
scipy==1.5.4
